Build Fixes Refactored TypeHelper into JavaType helper Added new RamlType helper for datatype utility methods Test Fixes
